feat(api): add v2 support with version-aware ui
